found ajax bug in master

+++ app/helpers/comments_helper.rb
@@ -14,7 +14,7 @@ module CommentsHelper
 
 
   def display_delete_link_comment(comment)
-    if validate_current_user
+    if validate_current_user && comment
       str = link_to("Delete", comment_path(comment), method: "delete", class: "col-md-1 pull-right")
 
       str.html_safe
